Rename game_file_info.php -> GameFileInfo.php
